refactor: split studio slice to explorer slice
