+++ src/store/studio/explorer/interfaces.ts
@@ -1,0 +1,8 @@
+import { ExplorerType } from '@/app/pages/Studio/constants';
+
+export interface ExplorerState {
+  width: number;
+  dragging: boolean;
+  explorer: ExplorerType;
+  currentExplorer: ExplorerType | null;
+}
