+++ src/store/studio/index.ts
@@ -1,3 +1,5 @@
 export * from './selectors';
 export * from './studioSlice';
 export { default } from './studioSlice';
+
+export * from './explorer';
