+++ src/store/studio/interfaces.ts
@@ -1,9 +1,3 @@
-import { ExplorerType } from '@/app/pages/Studio/constants';
-
 export interface StudioState {
-  width: number;
   flowId: string;
-  dragging: boolean;
-  explorer: ExplorerType;
-  currentExplorer: ExplorerType | null;
 }
