+++ src/store/studio/studioSlice.ts
@@ -1,15 +1,9 @@
 import { PayloadAction, createSlice } from '@reduxjs/toolkit';
 
-import { ExplorerType, MIN_EXPLORER_WIDTH } from '@/app/pages/Studio/constants';
-
 import type { StudioState } from './interfaces';
 
 const initialState: StudioState = {
   flowId: '',
-  dragging: false,
-  width: MIN_EXPLORER_WIDTH,
-  explorer: ExplorerType.FLOW,
-  currentExplorer: ExplorerType.FLOW,
 };
 
 // TODO: Cache previous width before closing
@@ -17,26 +11,6 @@ const studioSlice = createSlice({
   name: 'studio',
   initialState,
   reducers: {
-    setExplorer: (state, action: PayloadAction<ExplorerType | null>) => {
-      state.currentExplorer = action.payload;
-
-      if (action.payload) {
-        state.explorer = action.payload;
-      }
-    },
-    setDragging: (state, action: PayloadAction<boolean>) => {
-      state.dragging = action.payload;
-    },
-    setWidth: (state, action: PayloadAction<number>) => {
-      const width = action.payload;
-
-      if (width === 0) {
-        state.currentExplorer = null;
-      } else if (width > 0) {
-        state.width = width;
-        state.currentExplorer = state.explorer;
-      }
-    },
     setFlowId: (state, action: PayloadAction<string>) => {
       state.flowId = action.payload;
     },
@@ -45,5 +19,4 @@ const studioSlice = createSlice({
 
 export default studioSlice.reducer;
 
-export const { setDragging, setExplorer, setWidth, setFlowId } =
-  studioSlice.actions;
+export const { setFlowId } = studioSlice.actions;
